requirements.txt
Flask==2.1.1
openai==0.27.0
python-dotenv==0.19.2
gunicorn==20.1.0
fastapi
uvicorn
pydantic
sqlalchemy
databases[sqlite]
python-multipart
pytz